Create README.md
I'm new to coding but I've been and currently am using boot camps to become more familiar with several types of code. I'm also taking the free courses on edx to test my abilities and see if this is a good fit for me. I thoroughly enjoy learning code and testing my acquired knowledge to see what I can create. I would be happy to try and help anyone in creating any application, website or game. I'm new to this so if it's complex I probably wont be of much use, but if you're new like me I would like to help in any way that I can and hopefully together we can create something great!